testing flow log funcs
